First working version
Fixed a bunch of bugs, added better command line parsing.
Now properly encodes a .json preset into SysEx.
The encoded preset can be saved to a file or sent to the QuNeo.

The SysEx commands to load the preset in the QuNeo is also generated to
the resulting SysEx file. This makes QuNeo use the new presets
immediately.

Signed-off-by: Slinky <slinky@iki.fi>